3x_ui/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
py3xui